backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
qdrant-client==1.16.2
cohere==5.13.0
python-dotenv==1.0.1
openai-agents==0.6.4
pydantic==2.12.3